adding spells, pt.7 (FINAL) - adding explosion
Adding a spell that require a quite few changes to make it happen, the explosion. Which is a spell that deals damage in area, I had to create a new node (explo), import new sprites, and the spell is basically instantiating these nodes and checking for enemies in that area to deal damage, really interesting!